Separated-out ToDoSerializer from ToDoListSerializer
